Reflect changes from should.js v7.x.x
In version 7.0.0 the main file in the package.json of should.js changed
from './lib/should.js' to './index.js'. Due to the decreased directory
depth, the path-building routine for karma didn't point to a valid
location.

State lowest version of should.js that introduced the restructuring and
conditionally reduce the path's depth by 1.